requirements.txt
numpy==1.26.4
Pillow
black
pre-commit
pyright
pylint
pandas
scikit-learn
picsellia
ultralytics
pyyaml
types-PyYAML
mypy
